requirements.txt
pycryptodome==3.19.0
gt-telem==1.1.1
flask==3.0.0
flask-socketio==5.3.5
